refactor: view 관련 로직 view/ 로 이동
