+++ src/index.js
@@ -1,16 +1,16 @@
 import CarRace from "./domain/CarRace.js";
-import View from "./domain/View.js";
 import Car from "./domain/Car.js";
+import IO from "./view/IO.js";
 
 async function play() {
   // 자동차 이름 입력
-  const carNames = await View.getCarNamesFromUserInput();
+  const carNames = await IO.getCarNamesFromUserInput();
 
   // 자동차 이름으로부터 자동차 생성
   const cars = Car.createCarsFromCarNames(carNames);
 
   // 자동차 경주 횟수 입력
-  const totalRaceCount = await View.getTotalRaceCountFromUserInput();
+  const totalRaceCount = await IO.getTotalRaceCountFromUserInput();
 
   console.log("\n실행 결과");
 
@@ -21,7 +21,7 @@ async function play() {
   carRace.race();
 
   // 우승자 출력
-  View.printCarRaceWinners(carRace.winnerNames.join(", "));
+  IO.printCarRaceWinners(carRace.winnerNames.join(", "));
 }
 
 play();
